refactor: extract readBodyWithLimit helper
