+++ src/app/api/transactions/sync/route.ts
@@ -2,6 +2,7 @@ import { NextResponse } from "next/server"
 import { z } from "zod"
 import { verifyFirebaseToken } from "@/lib/server-auth"
 import { TransactionPayloadSchema } from "@/lib/transactions"
+import { readBodyWithLimit } from "@/lib/http"
 
 /**
  * Generic transaction syncing endpoint.
@@ -15,34 +16,6 @@ const bodySchema = z.object({
 })
 
 const MAX_BODY_SIZE = 1024 * 1024 // 1MB
-
-async function readBodyWithLimit(req: Request, limit: number) {
-  const contentLength = req.headers.get("content-length")
-  if (contentLength && Number(contentLength) > limit) {
-    return null
-  }
-
-  const reader = req.body?.getReader()
-  if (!reader) {
-    return ""
-  }
-  const decoder = new TextDecoder()
-  let total = 0
-  let result = ""
-  while (true) {
-    const { done, value } = await reader.read()
-    if (done) break
-    if (value) {
-      total += value.length
-      if (total > limit) {
-        return null
-      }
-      result += decoder.decode(value, { stream: true })
-    }
-  }
-  result += decoder.decode()
-  return result
-}
 
 export async function POST(req: Request) {
   try {
